python/requirements.txt
# Web scraping and data processing
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.3
numpy>=1.23.2,<2.0
aiohttp==3.11.15

# GitHub API
PyGithub==2.5.0
gitpython==3.1.43

# Text processing
markdown==3.7
pypandoc==1.14
python-magic==0.4.27
chardet==5.2.0

# Code parsing
tree-sitter==0.24.0
tree-sitter-python==0.23.6
tree-sitter-javascript==0.23.1
tree-sitter-typescript==0.23.2
tree-sitter-go==0.23.4
tree-sitter-rust==0.24.0
tree-sitter-java==0.23.5
tree-sitter-cpp==0.23.4

# Text chunking and processing
langchain==0.3.11
langchain-text-splitters==0.3.3
tiktoken==0.8.0

# Embedding models
sentence-transformers==3.3.1
transformers==4.53.2
torch
accelerate==1.2.1

# BGE Chinese model
FlagEmbedding==1.2.11

# Vector database
pymilvus==2.5.12
milvus-lite==2.5.1

# Search and retrieval
rank-bm25==0.2.2

# Reranking
flashrank==0.2.8

# Configuration and logging
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
loguru==0.7.2

# Utilities
tqdm==4.67.1
click==8.1.8
rich
httpx==0.28.1

# Development
pytest==8.3.4
black==24.10.0
isort==5.13.2
flake8==7.1.1 